# C2-2727: improve loadAuthenticationSubject performance